# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IRCUB API
  description: |
    Integrated Revenue Collection & Utility Billing Platform (POC).

    Authenticate with `POST /api/auth/login`. The SPA uses the HttpOnly
    `ircub_token` cookie. For API tools, send `X-IRCUB-Return-Token: 1` on login
    to receive a Bearer token, then pass `Authorization: Bearer <token>`.
    Most routes also require a permission slug (see role seeders).

    Mock external systems are under `/mock-api/*` (local/testing only).
  version: 1.0.0
  contact:
    name: Yusuf Mohamed Ahmed — Farsight Africa POC
servers:
  - url: http://127.0.0.1:8001
    description: Local artisan serve
  - url: https://ircub.waagefaal.so
    description: Hosted demo (when deployed)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
...
paths:
...
  /api/water-bills:
    get:
      tags: [Water]
      summary: List water bills
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
...
  /api/water-bills/{waterBill}:
    get:
      tags: [Water]
      summary: Water bill detail
      security:
        - bearerAuth: []
      parameters:
        - name: waterBill
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
...